delete modules directory
