+++ src/categories/helpers/categoryExists.ts
