+++ src/categories/types/category.ts
